layoutKey switch and set style.

+++ keyboard/layout/layout-file/symbol.layout.ts
@@ -137,7 +137,7 @@ const Symbolization = {
       ]
     ]
   },
-  'InitSet': 'Double'
+  'InitSet': 'Single'
 };
 
 export default Symbolization;
